python_scripts/requirements.txt
pymongo==4.1.0
openai==0.27.0
python-dotenv==0.21.0
